# bank-employees-salaries-api/resources/openapi.yaml
openapi: 3.0.1
info:
  title: API для работы с данными посещаемости, отделами и сотрудниками
  version: 1.0.0
  description: |
    Система формирования отчетов для AES.
    Включает компоненты для работы с моделями и DTO.
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
  - url: https://api.example.com/v1
    description: Продуктивный сервер
...
paths:
...
  /work-schedules:
    post:
      tags:
        - WorkSchedule
      summary: Добавить график в БД
      operationId: addSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'WorkScheduleDto.yaml#/components/schemas/WorkSchedule'
      responses:
        '200':
          description: График успешно добавлен
          content:
            application/json:
              schema:
                $ref: 'WorkScheduleDto.yaml#/components/schemas/WorkScheduleDto'

    get:
      tags:
        - WorkSchedule
      summary: Получить все графики в БД
      operationId: getSchedules
      responses:
        '200':
          description: Список графиков работы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'WorkScheduleDto.yaml#/components/schemas/WorkScheduleDto'

    delete:
      tags:
        - WorkSchedule
      summary: Удалить все графики из БД
      operationId: deleteSchedules
      responses:
        '204':
          description: Графики успешно удалены
  /work-schedules/{scheduleId}:
    patch:
      tags:
        - Work Schedule
      summary: Изменить график в БД
      operationId: updateSchedule
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'WorkScheduleDto.yaml#/components/schemas/WorkSchedule'
      responses:
        '200':
          description: График успешно обновлен
          content:
            application/json:
              schema:
                $ref: 'WorkScheduleDto.yaml#/components/schemas/WorkScheduleDto'

    delete:
      tags:
        - WorkSchedule
      summary: Удалить график в БД по id
      operationId: deleteScheduleById
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: График успешно удален

    get:
      tags:
        - Work Schedule
      summary: Получить график в БД по id
      operationId: getScheduleById
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '200':
          description: График работы
          content:
            application/json:
              schema:
                $ref: 'WorkScheduleDto.yaml#/components/schemas/WorkScheduleDto'